Trigger cloning Manuscript projects asynchronously
The first load of a manuscript for transcription is a costly operation,
as the entire project needs to be copied for the user.  Instead of
displaying a blank screen while this process is underway, present the
user with a loading screen, and only once the copying process is
complete do we redirect the user to T-PEN.

This also allows us to display a more elegant failure message when
something goes wrong during the copy process.

Conflicts:
	includes/tpenconnector.inc